[genAsm] typo in str translate

+++ compiler/symbol.cpp
@@ -426,7 +426,7 @@ string Var::getRawStr(){
 				}
 				else{
 				}
-				ss << "\"" << (int)strVal[i];
+				ss << "\"" << strVal[i];
 			}
 			else{
 				ss << strVal[i];
